Documentation

QuanticsGrids.DiscretizedGrid — Type

DiscretizedGrid{D}

A discretized grid structure for D-dimensional grids with variable resolution, supporting efficient conversion between quantics, grid indices, and original coordinates. A DiscretizedGrid instance is intended to undergird a quantics tensor train with a specific index structure, as defined in the indextable field. For example, say indextable is [[(:a, 1), (:b, 2)], [(:a, 2)], [(:b, 1), (:a, 3)]], then the corresponding tensor train has 3 tensor cores:

  a_1 b_2          a_2          b_1 a_3
   |   |            |            |   |
┌──┴───┴──┐    ┌────┴────┐    ┌──┴───┴──┐
│         │    │         │    │         │
│         │────│         │────│         │
│         │    │         │    │         │
└─────────┘    └─────────┘    └─────────┘

This object may be constructed with

julia> grid = DiscretizedGrid((:a, :b), [[(:a, 1), (:b, 2)], [(:a, 2)], [(:b, 1), (:a, 3)]])
DiscretizedGrid{2} with 8×4 = 32 grid points
├─ Variables: (a, b)
├─ Resolutions: (a: 3, b: 2)
├─ Domain: unit square [0, 1)²
├─ Grid spacing: (Δa = 0.125, Δb = 0.25)
└─ Tensor train: 3 sites (dimensions: 4-2-4)

and represents a 2^3 x 2^2 discretization of the unit square in the 2D plane (the x mark grid points):

   1.0  ┌───────────────────────────────┐
        │                               │
        │                               │
   0.75 x   x   x   x   x   x   x   x   │
        │                               │
        │                               │
b  0.5  x   x   x   x   x   x   x   x   │
        │                               │
        │                               │
   0.25 x   x   x   x   x   x   x   x   │
        │                               │
        │                               │
   0.0  x───x───x───x───x───x───x───x───┘
       0.0     0.25    0.5     0.75    1.0

                        a

If something other than a unit square is desired, lower_bound and upper_bound can be specified. Also, bases different than the default base 2 can be used, e.g. base=3 for a ternary grid.

In addition to the plain methods, there is a convenience layer for conversion from the original coordinates

julia> origcoord_to_grididx(grid; a=0.5, b=0.25)
(5, 2)

julia> origcoord_to_quantics(grid; a=0.5, b=0.25)
3-element Vector{Int64}:
 4
 1
 1

and also from grid indices

julia> grididx_to_origcoord(grid; a=5, b=2)
(0.5, 0.25)

julia> grididx_to_quantics(grid; a=5, b=2)
3-element Vector{Int64}:
 4
 1
 1

For a simpler grid, we can just supply the resolution in each dimension:

julia> boring_grid = DiscretizedGrid((3, 9))
DiscretizedGrid{2} with 8×512 = 4096 grid points
├─ Resolutions: (1: 3, 2: 9)
├─ Domain: unit square [0, 1)²
├─ Grid spacing: (Δ1 = 0.125, Δ2 = 0.001953125)
└─ Tensor train: 12 sites (uniform dimension 2)

In this case, variable names are automatically generated as 1, 2, etc.