Documentation

QuanticsGrids.DiscretizedGrid — Type

The DiscretizedGrid struct represents a grid for discretized continuous data. This is used for data that is originally continuous, but has been discretized for computational purposes. The grid contains values at specific, equally spaced points, which represent discrete approximations of the original continuous data.

QuanticsGrids.InherentDiscreteGrid — Type

The InherentDiscreteGrid struct represents a grid for inherently discrete data. The grid contains values at specific, equally spaced points, but these values do not represent discretized versions of continuous data. Instead, they represent individual data points that are inherently discrete. The linear size of the mesh is $base^R$, where $base$ defaults to 2.

QuanticsGrids.InherentDiscreteGrid — Method

Create a grid for inherently discrete data with origin at 1

QuanticsGrids._digit_at_index — Method

index, position and the result are one-based.

base: base index: The integer to look at. position=1: Specify the position of the digit to look at. 1 is the most significant (most left) digit. numdigits=8: Specify the number of digits in the number index.

QuanticsGrids.deinterleave_dimensions — Method

deinterleave_dimensions(digitlist, d)

Reverses the interleaving of bits, i.e. yields digitlists for each dimension from a long interleaved digitlist. Inverse of interleave_dimensions.

QuanticsGrids.fuse_dimensions — Method

fuse_dimensions(digitlists...)

Fuse d digitlists that represent a quantics index into a digitlist where each bit has dimension base^d. This fuses legs for different dimensions that have equal length scale (see QTCI paper).

Inverse of unfuse_dimensions.

QuanticsGrids.fused_to_interleaved — Method

Converrt a fused qunaitcs representation to an unfused quantics representation

QuanticsGrids.grididx_to_origcoord — Method

grid index => original coordinate

QuanticsGrids.grididx_to_origcoord — Method

grid index => original coordinate

QuanticsGrids.grididx_to_quantics — Method

grid index => quantics

QuanticsGrids.grididx_to_quantics — Method

grid index => quantics

QuanticsGrids.index_to_quantics! — Method

function index_to_quantics!(digitlist, index::Integer; base::Integer=2)

digitlist base-b representation (1d vector)

base base for quantics (default: 2)

QuanticsGrids.index_to_quantics — Method

index_to_quantics(index::Integer; numdigits=8, base::Integer=2)

Does the same as index_to_quantics! but returns a new vector.

QuanticsGrids.index_to_quantics_fused! — Method

index_to_quantics_fused!(digitlist::AbstractVector{<:Integer}, index::NTuple{D,<:Integer}; base::Integer=2) where {D}

Does the opposite of quantics_to_index_fused

D

QuanticsGrids.interleave_dimensions — Method

interleave_dimensions(digitlists...)

Interleaves the indices of all digitlists into one long digitlist. Use this for quantics representation of multidimensional objects without fusing indices. Inverse of deinterleave_dimensions.

QuanticsGrids.interleaved_to_fused — Method

Converrt a unfused qunaitcs representation to an fused quantics representation

QuanticsGrids.origcoord_to_grididx — Method

Convert a coordinate in the original coordinate system to the corresponding grid index

QuanticsGrids.origcoord_to_grididx — Method

Convert a coordinate in the original coordinate system to the corresponding grid index

QuanticsGrids.origcoord_to_quantics — Method

original coordinate => quantics

QuanticsGrids.quantics_to_grididx — Method

quantics => grid index

QuanticsGrids.quantics_to_index — Method

function quantics_to_index(
    digitlist::AbstractVector{<:Integer};
    base::Val{B}=Val(2), dims::Val{d}=Val(1),
    unfoldingscheme::Symbol=:fused
)::NTuple{d,Int} where {B, d}

Convert a d-dimensional index from fused quantics representation to d Integers.

base base for quantics (default: 2)

d number of dimensions

digitlist base-b representation

unfoldingscheme Unfolding scheme (fused or interleaved)

QuanticsGrids.quantics_to_index_fused — Method

function quantics_to_index_fused(
    digitlist::AbstractVector{<:Integer};
    base::Val{B}=Val(2), dims::Val{d}=Val(1)
)::NTuple{d,Int} where {B, d}

Convert a d-dimensional index from fused quantics representation to d Integers.

base base for quantics (default: 2)

d number of dimensions

digitlist base-b representation

See also quantics_to_index.

QuanticsGrids.quantics_to_origcoord — Method

quantics => original coordinate system

QuanticsGrids.quanticsfunction — Method

Make a wrapper function that takes a bitlist as input

QuanticsGrids.unfuse_dimensions — Method

unfuse_dimensions([base=Val(2)], digitlist, d)

Unfuse up a fused digitlist with bits of dimension base^d into d digitlists where each bit has dimension base. Inverse of fuse_dimensions.